shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "aio-pika>=9.6.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.0",
    "redis>=7.2.0",
    "sqlalchemy>=2.0.46",
]

[tool.uv]
package = false
